[feat] init view stateholder
todayView의 stateholder을 init
